staging: unisys: Re-enable interrupts after we have done the work
Instead of rescheduling the work queue after we have called the
channel interrupt, start it when the driver has finished processing
the queue.

This patch introduces the visorbus_rearm_channel_interrupts function
that must get called when the driver decides that it is done processing
its queue.

Visorinput, visorhba, and visornic were all updated to call the new
function.

Signed-off-by: David Kershner <david.kershner@unisys.com>
Signed-off-by: Benjamin Romer <benjamin.romer@unisys.com>