staging: unisys: visornic: prevent hang doing 'modprobe -r visornic'
A stray+extraneous 'netif_napi_add()' that we were doing in visornic_probe()
was causing havoc when we got into visornic_remove(), called during
'modprobe -r visornic'.  The symptom was a processor busy-wait loop on the
modprobe process, which '/proc/<pid>/stack' would show looping doing napi
things.

Presumably the stray line got there as a result of some merging snafoo, and
has been deleted to fix the problem.  With this patch 'modprobe -r visornic'
and a subsequent 'modprobe visornic' both complete successfully, and result
in an operational network.

githubissue: #64
Signed-off-by: Tim Sell <Timothy.Sell@unisys.com>
Signed-off-by: David Kershner <david.kershner@unisys.com>