staging: unisys: add simple versioning to visor drivers
Previously, the visor drivers made use of a common version.h which
used an awkward define scheme to generate a version string.  This
scheme made sense historically, but no longer does today.  It
has been removed in favor of a simple versioning scheme in which the 4
visor drivers, visorbus, visorhba, visorinput, and visornic can define
their own unique versions.

We will use a simple version number in the form:
<major-ver>.<year>.<month>.<day>

Signed-off-by: Jon Frisch <jon.frisch@unisys.com>
Signed-off-by: David Kershner <david.kershner@unisys.com>