staging: unisys: visorbus: Relocate logic of visorbus_free_irq()
Relocates the logic of visorbus_free_irq() to both an earlier function,
and a brand-new function. Specifically, the code that disables the irq is
relocated to visorbus_disable_channel_interrupts() while the code that
unregisters the irq is placed into a new function:
visorbus_unregister_for_channel_interrupts(). This new function is then
invoked in visornic_remove(), directly after the invocation to
visorbus_disable_channel_interrupts().

Signed-off-by: David Binder <david.binder@unisys.com>