staging: unisys: visorinput: fix crashes involving visor_device.timer
The bug fixed by this patch would cause a

    BUG at kernel/timer.c:933!

approximately 1 in every 10 boots of RHEL 7.2.

Problems were occurring because a rogue call of interrupt_work() was
occurring AFTER interrupts had been disabled,
(visorbus_disable_channel_interrupts()), due to interrupt_work() being
invoked by work scheduled out of visorinput_channel_interrupt().

interrupt_work() does this before returning:

    visorbus_rearm_channel_interrupts(devdata->dev);

which would cause us to do a mod_timer() on visor_device.timer.  At this
point we would be in a situation where the code THOUGHT the timer should
be inactive (because interrupts had been disabled), but in fact the timer
was active (pending).  If at this point a device open() occurred:
  --> visorinput_open()
      --> visorbus_enable_channel_interrupts()
          --> dev_start_periodic_work()
              --> add_timer()

The add_timer() would blow up because the timer was active (pending), with
a stacktrace looking like:

    kernel BUG at kernel/timer.c:933!
    visorbus_enable_channel_interrupts+0x15a/0x1e0 [visorbus]
    visorinput_open+0x68/0xb0 [visorinput]
    input_open_device+0x79/0xa0
    evdev_open+0x191/0x1d0
    chrdev_open+0xa1/0x1e0
    do_entry_open+0x1a7/0x2e0

The patch guarantees that we can't ever get into interrupt_work() after
interrupts have been disabled, by making sure that cancel_work_sync()
happens after EVERY call to visorbus_disable_channel_interrupts().

Note that there was already a hint of the solution within
visorinput_remove() immediately after it calls
visorbus_disable_channel_interrupts():

    cancel_work_sync(&devdata->interrupt_work);

cancel_work_sync() can block, but that's ok, because every place we call
visorbus_disable_channel_interrupts(), we are prepared to block.

= Testing =

With this patch running on a RHEL 7.2 guest, I was able to boot 40 times
in a row without a failure.  The most successful boots we were able to get
withOUT the patch was 11.  The fault never occurred with the patch.

Also ran this script, and ensured that the keyboard and mouse in the
partition desktop were both operational during the 15-second periods where
visorinput was loaded:

    #!/bin/sh
    set -ex

    i=0
    while [ $i -lt 30 ]
    do
        rmmod visorinput
        sleep 15
        modprobe visorinput
        sleep 15
        let i=i+1
    done

Signed-off-by: Tim Sell <Timothy.Sell@unisys.com>